# backend/docs/swagger.yaml
basePath: /v1/api/radar-hub-manager
definitions:
...
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Schedule:
    properties:
      commander:
        description: ====== Thông tin kíp trực cho khung giờ này ======
        type: string
      created_at:
        type: integer
      crew:
        description: Danh sách kíp trực
        type: string
      end_hhmm:
        description: '"0330"'
        type: string
      id:
        type: integer
      phone:
        description: Số liên lạc
        type: string
      start_hhmm:
        description: '"0130"'
        type: string
      station_id:
        type: integer
      updated_at:
        type: integer
    type: object
...
  internal_handlers.CreateScheduleRequest:
    properties:
      commander:
        type: string
      crew:
        type: string
      end_hhmm:
        type: string
      phone:
        type: string
      start_hhmm:
        type: string
    required:
    - end_hhmm
    - start_hhmm
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.UpdateScheduleRequest:
    properties:
      commander:
        type: string
      crew:
        type: string
      end_hhmm:
        type: string
      phone:
        type: string
      start_hhmm:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Radar Hub Manager system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Radar Hub Manager API
  version: "1.0"
paths:
...
  /station-schedules/station/{station_id}:
    get:
      description: Get all schedules for a specific station (Operator only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Schedule'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List schedules for a station
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Create a new schedule for a station (Operator only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Schedule data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new schedule
      tags:
      - schedules
  /station-schedules/station/{station_id}/{schedule_id}:
    delete:
      description: Delete a schedule by station ID and schedule ID (Operator only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a schedule
      tags:
      - schedules
    get:
      description: Get a specific schedule by station ID and schedule ID (Operator
        only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a schedule by ID
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: Update an existing schedule by station ID and schedule ID (Operator
        only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      - description: Updated schedule data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a schedule
      tags:
      - schedules
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"